// frontend/packages/hydrangean-diva/src/api/models/HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HydrangeanDiva
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto
 */
export interface HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto {
    /**
     * 
     * @type {string}
     * @memberof HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto
     */
    type: string;
    /**
     * 
     * @type {string}
     * @memberof HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto
     */
    videoId: string;
    /**
     * 
     * @type {string}
     * @memberof HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto
     */
    title: string;
}

/**
 * Check if a given object implements the HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto interface.
 */
export function instanceOfHydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto(value: object): value is HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('url' in value) || value['url'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('videoId' in value) || value['videoId'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    return true;
}

export function HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDtoFromJSON(json: any): HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto {
    return HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDtoFromJSONTyped(json, false);
}

export function HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'url': json['url'],
        'type': json['type'],
        'videoId': json['videoId'],
        'title': json['title'],
    };
}

export function HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDtoToJSON(json: any): HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto {
    return HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDtoToJSONTyped(json, false);
}

export function HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDtoToJSONTyped(value?: HydrangeanDivaMediaPlayerContractsPlaylistsDtosTrackDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'url': value['url'],
        'type': value['type'],
        'videoId': value['videoId'],
        'title': value['title'],
    };
}